# Enhancements to bignums (including a big speedup to `bignum2double` and `exact->inexact`)