@jpellegrini 2023-04-04
By the way... Guile also does have an optimized version of `exact->inexact`, but they use the low-level GMP functions that work on the internal GMP representation for numbers (the functions that deal with number "limbs") -- I thought this would be a bit too complex and didn't go that way.